feat: improve isLoggedIn This corresponds to this commit in the original repo: * https://github.com/n0madic/twitter-scraper/commit/2b607217fd1165007e9e2f52fea03b3e0df4e2e3
